+++ sparse/coo/__init__.py
@@ -1,5 +1,5 @@
 from .core import COO, as_coo
 from .umath import elemwise
 from .common import (tensordot, dot, concatenate, stack, triu, tril, where,
-                     nansum, nanprod, nanmin, nanmax, nanreduce, roll, eye,
-                     full, full_like, zeros, zeros_like, ones, ones_like)
+                     nansum, nanmean, nanprod, nanmin, nanmax, nanreduce, roll,
+                     eye, full, full_like, zeros, zeros_like, ones, ones_like)